backend/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
flask>=2.0.1
numpy>=1.21.0
Pillow>=9.0.0
scikit-learn>=1.0.0
seaborn>=0.11.0
matplotlib>=3.4.0 